パスワードの再入力の追加

+++ view/js/register.js
@@ -24,25 +24,27 @@ function inputRegistrationButton(){
 	var param = [ $('#username').val(), $('#password').val(), fileName, $('#mail').val() ];
 	data.append('list', param);
 
-    $.ajax({
-    	url         : '../../api/controller.php',
-    	type        : 'POST',
-    	dataType    : 'json',
-    	processData : false,
-    	contentType : false,
-    	data        :  data,
-    	timeout     :  1000,
-    }).done(function(data, dataType){
-    	if(data[0].result == 'success') {
-    		location.href = "../html/mypage.html";
-            sessionStorage.setItem('userId', data[0].id);
-    	} else {
-    		alert(data);
-    	}
+	if(window.confirm('この内容で登録しますか？')){
+		$.ajax({
+	    	url         : '../../api/controller.php',
+	    	type        : 'POST',
+	    	dataType    : 'json',
+	    	processData : false,
+	    	contentType : false,
+	    	data        :  data,
+	    	timeout     :  1000,
+	    }).done(function(data, dataType){
+	    	if(data[0].result == 'success') {
+	    		location.href = "../html/mypage.html";
+	            sessionStorage.setItem('userId', data[0].id);
+	    	} else {
+	    		alert(data);
+	    	}
 
-    }).fail(function(){
-    	alert('Nodata');
-    });
+	    }).fail(function(){
+	    	alert('Nodata');
+	    });
+	}
 }
 
 // 一覧画面へ移動
@@ -106,7 +108,9 @@ function checkValidation(_file){
 	var lg = _file[0].files.length;
 	var item = _file[0].files;
 	var name = $('[name="user"]').val();
-	var password = $('[name="password"]').val();
+	// var password = $('[name="password"]').val();
+	var password = $('#password').val();
+	var password2 = $('#password2').val();
 	var mail = $('[name="mail"]').val();
 	var string = "";
 
@@ -140,11 +144,12 @@ function checkValidation(_file){
 		}
 	}
 
-	if(password == ""){
+	if(password == "" || password2 == ""){
 		string = string + 'パスワード入力してください。';
 	}else if(password.length < 4){
 		string = string +'パスワードは最低4文字必要です。';
-	}
+	}else if(password != password2)
+		string = string +'パスワードが一致しませんでした。';
 
 
 	if(string != ""){
